Changed label setting to happen on resizing instead of on rendering
