+++ app/digital/ts/models/ioobjects/other/Decoder.ts
@@ -22,6 +22,7 @@ export class Decoder extends DigitalComponent {
 
         // activate 0th port for initial state
         super.activate(true, 0);
+        this.setBinaryLabels();
     }
 
     public activate(): void {
@@ -42,9 +43,21 @@ export class Decoder extends DigitalComponent {
         const width = Math.max(2 * padding + val * 8.342285, DEFAULT_SIZE);
         this.transform.setSize(V(width, DEFAULT_SIZE/2*Math.pow(2, val)));
         super.setInputPortCount(val);
+        this.setBinaryLabels();
     }
 
     public getDisplayName(): string {
         return "Decoder";
+    }
+
+    public setBinaryLabels(): void {
+        const ports = this.getOutputPorts();
+        const digitCount = Math.log2(ports.length);
+        let numStr = "0".repeat(digitCount);
+
+        for (let i = 0; i < Math.pow(2, digitCount); i++) {
+            ports[i].setName(numStr);
+            numStr = (numStr.substr(0, numStr.lastIndexOf("0")) + "1").padEnd(digitCount, "0");
+        }
     }
 }
