+++ app/digital/ts/models/ioobjects/other/Demultiplexer.ts
@@ -1,6 +1,7 @@
 import {ClampedValue} from "math/ClampedValue";
 
 import {OutputPort} from "../../ports/OutputPort";
+import {Port} from "../../../../../core/ts/models/ports/Port";
 import {MuxPositioner,
         MuxSinglePortPositioner} from "../../ports/positioners/MuxPositioners";
 
@@ -34,5 +35,9 @@ export class Demultiplexer extends Mux {
 
     public getDisplayName(): string {
         return "Demultiplexer";
+    }
+
+    public getLabeledPorts(): Port[] {
+        return this.getOutputPorts();
     }
 }
