+++ app/digital/ts/models/ioobjects/other/Mux.ts
@@ -26,6 +26,7 @@ export abstract class Mux extends DigitalComponent {
         this.selects = new PortSet<InputPort>(this, new ClampedValue(2, 1, 8), new MuxSelectPositioner(), InputPort);
 
         this.setSelectPortCount(2);
+        this.setBinaryLabels();
     }
 
     public setSelectPortCount(val: number): void {
@@ -68,4 +69,16 @@ export abstract class Mux extends DigitalComponent {
         return super.getPorts().concat(this.getSelectPorts());
     }
 
+    public abstract getLabeledPorts(): Port[];
+
+    public setBinaryLabels(): void {
+        const ports = this.getLabeledPorts();
+        const digitCount = Math.log2(ports.length);
+        let numStr = "0".repeat(digitCount);
+
+        for (let i = 0; i < Math.pow(2, digitCount); i++) {
+            ports[i].setName(numStr);
+            numStr = (numStr.substr(0, numStr.lastIndexOf("0")) + "1").padEnd(digitCount, "0");
+        }
+    }
 }
